requirements.txt
django
django-rosetta
django-simple-pwa
django-allauth
SpeechRecognition
pyyaml
h5py
matplotlib
torchvision
torchsummary
tensorflow
pyAudio
pillow
geopy
rfc3339
rosetta
scipy
pandas
nltk
mysql
torch
mysql-connector
google-generativeai
